Rename to hawaii-workspace
The new name reflects better what is contained here.
We have a screenshot application which is definitely one of the base
applications and the PolicyKit agent that is more a workspace
component.

Base applications are option (not required to run a full blown Hawaii
session) but they are still important to get the full user experience,
so we want them to be installed along other stuff and rename the repository
to reflect the central role it has gained now.